@@ -1,31 +1,60 @@
 package com.ensat.entities;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Version;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+
 import java.math.BigDecimal;
 
 /**
  * Product entity.
  */
 @Entity
-@Getter
-@Setter
-@NoArgsConstructor
+
 public class Product {
-
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Version
     private Integer version;
 
     private String productId;
     private String name;
-    private BigDecimal price;
+
+    public Integer getVersion() {
+        return version;
     }
 
+    public void setVersion(Integer version) {
+        this.version = version;
+    }
+
+    public String getProductId() {
+        return productId;
+    }
+
+    public void setProductId(String productId) {
+        this.productId = productId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
+
+    private BigDecimal price;
+
 }
+
