@@ -4,15 +4,7 @@ import com.ensat.entities.Product;
 import com.ensat.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-
-
+import org.springframework.web.bind.annotation.*;
 
 
 /**
@@ -79,7 +71,7 @@ public class ProductController {
     @RequestMapping(value = "product", method = RequestMethod.POST)
     public String saveProduct(Product product) {
         productService.saveProduct(product);
-        return "redirect:/product/" + product.getId();
+        return "redirect:/product/" + product.getProductId();
     }
 
     /**
